refactor: move Achieve class to new achieve.py
