+++ gmatpyplus/command/__init__.py
@@ -1,3 +1,4 @@
+from .achieve import Achieve
 from .begin_finite_burn import BeginFiniteBurn
 from .begin_mission_sequence import BeginMissionSequence
 from .branch_command import *
